move useStyles to BodyStyles.js for code split CSS
